Adding Makefile.  pyproject.toml is broken?

+++ Makefile
@@ -1,8 +1,12 @@
 
 
 environment:
-	rm -fr ./env
+	pwd
+	ls
+	if [ -e ./env ]; then rm -fr ./env; fi
+	if [ -e ./poetry.lock ]; then rm ./poetry.lock; fi
 	python -m venv env
 	. env/bin/activate
+	pip install poetry
 	poetry install
